Extend getUsers to return the e-mail addresses if available
When searching the sharees only the name of the sharee and its type was
returned. Now, if the sharee is a user, its e-mail address is also
returned (if it is set for that user); this will be used by the
front-end to provide more actions without needing to query the server
again for each matched user.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>