Fix leaking e-mail addresses
There is no need to send the e-mail address to the client in order to
resend the e-mail notification; simply knowing whether the user has an
e-mail address or not is enough. Due to this, and to prevent leaking
e-mail addresses set as private in the user profile, the "emailAddress"
field sent is replaced by just "hasEmailAddress".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>