Add support for touch events to jQuery UI
jQuery UI does not support touch events out of the box, so the features
that depend on jQuery UI, like dragging files in the Files app, are not
available on touch screens.

This commit imports the latest version of jQuery UI Touch Punch, which
adds support to touch events by mapping them to their mouse counterparts
and requires no changes to the existing code.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>